core: vm: sqlvm: ast: put source code tokens on AST nodes
In order to construct an error message which can be easily understood,
it is sometimes required to reference source code tokens corresponding
to nodes. It should not increase memory usage too much since all byte
slices share the same underlying array.

Drop UpdatePosition because it is never used. It is unclear that how it
can be used and it becomes obsolete after adding this new field.