core: vm: sqlvm: errors: add severity field
In order to report messages other than errors, we are going to reuse
the error struct to store non-error messages, such as warnings and
notes. We add a new field to distinguish between error and non-error
messages. This field is not a part of SQLVM ABI, so non-error messages
should never be returned to the caller of the contract.

The zero value of severity is 'error'. Code which doesn't set this field
should keep working because 'error' was the only use case.